Implemented new flags, bug fixes with scanning/validation

+++ pipeline/init.py
@@ -6,18 +6,16 @@ __copyright__ = "Copyright 2023 United Kingdom Research and Innovation"
 import os
 import json
 import logging
+import glob
 
 config = {
     'proj_code': None,
-    'workdir': None,
-    'proj_dir':None,
     'pattern': None,
     'update': None,
     'remove': None
 }
 
 levels = [
-    logging.ERROR,
     logging.WARN,
     logging.INFO,
     logging.DEBUG
@@ -25,7 +23,7 @@ levels = [
 
 def init_logger(verbose, mode, name):
     """Logger object init and configure with formatting"""
-    verbose = min(verbose, len(levels-1))
+    verbose = min(verbose, len(levels)-1)
 
     logger = logging.getLogger(name)
     logger.setLevel(levels[verbose])
@@ -62,7 +60,22 @@ def get_removals(logger):
     return valsarr
 
 def get_proj_code(path, prefix=''):
-    return path.replace(prefix,'').replace('/','_')
+    parts = path.replace(prefix,'').replace('/','_').split('_')
+    if '*.' in parts[-1]:
+        parts = parts[:-2]
+    return '_'.join(parts)
+    
+
+def make_filelist(pattern, proj_dir, logger):
+    """Create list of files associated with this project"""
+    logger.debug(f'Making list of files for project {proj_dir.split("/")[-1]}')
+
+    if os.path.isdir(proj_dir):
+        os.system(f'ls {pattern} > {proj_dir}/allfiles.txt')
+    else:
+        logger.error(f'Project Directory not located - {proj_dir}')
+        return None
+    return True
 
 def load_from_input_file(args, logger):
     """Configure project directory and base config from input file"""
@@ -80,37 +93,75 @@ def load_from_input_file(args, logger):
         logger.error(f'Input file {args.input} does not exist')
         return None
 
-def text_file_to_csv(args, logger):
+def text_file_to_csv(args, logger, prefix=None):
     """Convert text file list of patterns to a csv for a set of projects"""
     logger.debug('Converting text file to csv')
 
     if not os.path.isdir(args.workdir):
-        os.path.makedirs(args.workdir)
+        if args.dryrun:
+            logger.debug(f'DRYRUN: Skip making workdir {args.workdir}')
+        else:
+            os.makedirs(args.workdir)
     if args.groupdir and not os.path.isdir(args.groupdir):
-        os.path.makedirs(args.groupdir)
+        if args.dryrun:
+            logger.debug(f'DRYRUN: Skip making groupdir {args.groupdir}')
+        else:
+            os.makedirs(args.groupdir)
 
-    new_inputfile = f'{args.groupdir}/filelists/{args.group}.txt'
-    prefix = '' # Remove from path for proj_code
-    os.system(f'cp {args.input} {new_inputfile}')
+    new_inputfile = f'{args.workdir}/groups/filelists/{args.groupID}.txt'
+
+    if new_inputfile != args.input:
+        if args.dryrun:
+            logger.debug(f'DRYRUN: Skip copying input file {args.input} to {new_inputfile}')
+        else:
+            os.system(f'cp {args.input} {new_inputfile}')
 
     with open(new_inputfile) as f:
         datasets = [r.strip() for r in f.readlines()]
 
-    records = ''
-    for ds in datasets:
-        proj_code = get_proj_code(ds, prefix=prefix)
-        proj_dir  = f'{args.workdir}/in_progress/{args.group}/{proj_code}'
-        pattern   = f'{os.path.realpath(ds)}/*.nc'
-        records  += f'{proj_code},{args.workdir},{proj_dir},{pattern},,\n'
+    if not os.path.isfile(f'{args.groupdir}/datasets.csv') or args.forceful:
+        records = ''
+        logger.info('Creating filesets for each dataset')
+        for index, ds in enumerate(datasets):
+            skip = False
 
-    with open(f'{args.groupdir}/datasets.csv','w') as f:
-        f.write(records)
+            pattern = str(ds)
+            if not (pattern.endswith('.nc') or pattern.endswith('.tif')):
+                logger.debug('Identifying extension')
+                fileset = [r.split('.')[-1] for r in glob.glob(f'{pattern}/*')]
+                if len(set(fileset)) > 1:
+                    logger.error(f'File type not specified for {pattern} - found multiple ')
+                    skip = True
+                elif len(set(fileset)) == 0:
+                    skip = True
+                else:
+                    extension = list(set(fileset))[0]
+                    pattern = f'{pattern}/*.{extension}'
+                logger.debug(f'Found .{extension} common type')
+
+            if not skip:
+                proj_code = get_proj_code(ds, prefix=prefix)
+                logger.debug(f'Assembled project code: {proj_code}')
+                proj_dir  = f'{args.workdir}/in_progress/{args.groupID}/{proj_code}'
+                if 'latest' in pattern:
+                    pattern = pattern.replace('latest', os.readlink(pattern))
+
+                records  += f'{proj_code},{args.workdir},{proj_dir},{pattern},,\n'
+                logger.debug(f'Added entry and created fileset for {index+1}/{len(datasets)}')
+        if args.dryrun:
+            logger.debug(f'DRYRUN: Skip creating csv file {args.groupdir}/datasets.csv')    
+        else:        
+            with open(f'{args.groupdir}/datasets.csv','w') as f:
+                f.write(records)
+    else:
+        logger.warn(f'Using existing csv file at {args.groupdir}/datasets.csv')
+    return True
     
     # Output completed csv setup part
 
 def make_dirs(args, logger):
     """Set up directory structure for working directory"""
-    logger.debug('Creating project directories')
+    logger.info('Creating project directories')
 
     # Open csv and gather data
     with open(f'{args.groupdir}/datasets.csv') as f:
@@ -119,40 +170,76 @@ def make_dirs(args, logger):
     # Map dataset parameters from csv to config JSON
     params     = list(config.keys())
     proj_codes = list(datasets.keys())
-    for dskey in proj_codes:
-        cfg = dict(config)
-        ds  = datasets[dskey]
 
-        cfg[params[0]] = dskey # Set project code
+    if proj_codes[0] == 'Project Code':
+        proj_codes = proj_codes[1:]
+    
+    for index, proj_code in enumerate(proj_codes):
+        cfg_values = dict(config)      # Ensure no linking
+        ds_values  = datasets[proj_code]
+
+        logger.info(f'Creating directories/filelists for {index+1}/{len(proj_codes)}')
+
+        cfg_values[params[0]] = proj_code
         # Set all other parameters
-        for x, p in enumerate(params[1:]):
-            cfg[p] = ds[x]
+        if len(params) == len(ds_values):
+            for x, p in enumerate(params[1:]):
+                cfg_values[p] = ds_values[x]
+        else:
+            logger.warning(f'Project code {index}:{proj_code} from {args.groupID} does not have correct number of fields.')
+            logger.warning(f'Fields specified must be {list(params.keys())}')
+
+        if 'latest' in pattern:
+            pattern = pattern.replace('latest', os.readlink(pattern))
+
+        proj_dir = cfg_values['proj_dir']
 
         # Save config file
-        if not os.path.isdir(cfg['proj_dir']):
-            os.makedirs(cfg['proj_dir'])
-    
-            with open(f'{cfg["proj_dir"]}/base-cfg.json','w') as f:
-                f.write(json.dumps(cfg))
-        
+        if not os.path.isdir(proj_dir):
+            if args.dryrun:
+                logger.debug(f'DRYRUN: Skip making Directories to {proj_dir}')
+            else:
+                os.makedirs(proj_dir)
         else:
-            logger.warn(f'{cfg["proj_code"]} already exists - skipping')
+            if not args.forceful:
+                logger.warn(f'{proj_code} directory already exists')
 
-    logger.info(f'Exported {len(proj_codes)} dataset config files')
+        status = make_filelist(pattern, proj_dir, logger)
+        if not status:
+            logger.error(f'Issue creating filelist for {proj_code}')
+        else:
+            base_file = f'{proj_dir}/base-cfg.json'
 
-    with open(f'{args.groupdir}/proj_codes_1.txt','w') as f:
-        f.write('\n'.join(proj_codes))
+            if not os.path.isfile(base_file) or args.forceful:
+                if args.dryrun:
+                    logger.debug(f'DRYRUN: Skip writing base file {base_file}')
+                else:
+                    with open(base_file,'w') as f:
+                        f.write(json.dumps(cfg_values))
+            else:
+                logger.warn(f'{base_file} already exists - skipping')
 
-    logger.info(f'Written as group ID:{args.groupID}')
+    logger.info(f'Exporting {len(proj_codes)} dataset config files')
+
+    if args.dryrun:
+        logger.debug(f'DRYRUN: Skip writing {len(proj_codes)} project codes list {args.groupdir}/proj_codes_1.txt')
+    else:
+        with open(f'{args.groupdir}/proj_codes_1.txt','w') as f:
+            f.write('\n'.join(proj_codes))
+
+    logger.info(f'Written as group ID: {args.groupID}')
 
 def init_config(args):
     """Main configuration script, load configurations from input sources"""
 
     logger = init_logger(args.verbose, args.mode, 'init')
+    logger.info('Starting initialisation')
 
     groupID = None
     if hasattr(args, 'groupID'):
         groupID = getattr(args,'groupID')
+    if hasattr(args,'group'):
+        groupID = getattr(args,'group')
 
     if groupID:
         logger.debug('Starting group initialisation')
@@ -161,11 +248,14 @@ def init_config(args):
             return None
         
         if '.txt' in args.input:
-            text_file_to_csv(args, logger) # Includes creating csv
+            logger.debug('Converting text file to csv')
+            status = text_file_to_csv(args, logger) # Includes creating csv
+            if not status:
+                return None
         elif '.csv' in args.input:
+            logger.debug('Ingesting csv file')
             new_csv = f'{args.groupdir}/datasets.csv'
             os.system(f'cp {args.input} {new_csv}')
-
         make_dirs(args, logger)
 
     else:
